Emily Jupyter Hangman Start
Here is the work I did for the jupyter app. I found a dictionary of words for us and uploaded it here (words.txt). Download it onto your computer and put it in the same file as the ipynb and it should open easily since I already did the code. here is what I got done and what I did not manage to figure out based on the tasks he gave us (X means completed):
1. Ask if they want to play a game, if not then return X
2. Generate a random word (w) X Also initialize the counter variable which keeps track of their remaining guesses and initialize a variable to keep track of the letters they've guessed already. **I was not able to figure this out in total. I was able to keep track of guesses and end the game when they ran out, but not count them**
3. Ask the user for a letter and add it to the list of guessed letters X
4. If the letter has already been guessed or is not in the word, add one to the counter and print a message saying they didn't guess a new letter X
5. Otherwise, congratulate them and print the letters in the word, but only showing the letters they have guessed. **could not figure this out completely**
6. If they have guessed all the letters or they have used up all of their guesses end the game **a few bugs here, you will see in the file**
7. Otherwise continue the game X

Let me know if you have any questions, I think Pam was going after me (?) so we can chat on Zoom if you can't follow what I was doing