Bio's changes
I have added the heading, buttons, pictures, and fixed the formatting of how the bios are accessed. Pictures are from Google to keep, I am not too sure if that might change how things run in other computers. Please let me know if you have any trouble accessing it or viewing the pictures.